requirements.txt
Django==5.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.2.0
shippo==3.1.0
whitenoise==6.6.0
gunicorn==21.2.0
django-cors-headers==4.3.1
django-filter==23.5
